move tls class in a separate module
